requirements.txt
scipy==1.2.1
numpy==1.16.2
Keras==2.1.5
matplotlib==3.1.2
opencv_python==4.1.2.30
tensorflow==1.13.2
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0
